move SDL shims to a new DFSDL module
