+++ library/CMakeLists.txt
@@ -124,6 +124,7 @@ set(MODULE_HEADERS
     include/modules/Buildings.h
     include/modules/Burrows.h
     include/modules/Constructions.h
+    include/modules/DFSDL.h
     include/modules/Designations.h
     include/modules/EventManager.h
     include/modules/Filesystem.h
@@ -151,6 +152,7 @@ set(MODULE_SOURCES
     modules/Buildings.cpp
     modules/Burrows.cpp
     modules/Constructions.cpp
+    modules/DFSDL.cpp
     modules/Designations.cpp
     modules/EventManager.cpp
     modules/Filesystem.cpp
